+++ docs/RELEASE-NOTES.md
@@ -22,6 +22,7 @@
   - compiled with Scala 2.13.2
   - requires `scala-library-2.13.2.jar`
   - its artifact identifiers include the Scala version as suffix (`_2.13`)
+  - use of multiple tables is deprecated and they are no longer supported in newer versions
   bug_fixes :
   - respects the order of the defined URI prefixes when they are expanded
   new_writers :