+++ docs/RELEASE-NOTES.md
@@ -2,8 +2,8 @@
 - - version: v1.1.0
   - date: (*unreleased*)
   - new_features:
-    - compiled with Scala 2.13.2
-    - requires `scala-library-2.13.2.jar`
+    - compiled with Scala 2.13.3
+    - requires `scala-library-2.13.3.jar`
     - its artifact identifiers include the Scala version as suffix (`_2.13`)
   - deprecated:
     - use of multiple tables is deprecated and they are no longer supported in newer